chore: refactored cli commands to be defined in each service module
