+++ src/mightstone/cli/__main__.py
@@ -1,0 +1,3 @@
+from mightstone.cli import cli
+
+cli()
